added terminal support with integration to quickfix list
Open up a terminal split with <leader-t>
Replace current buffer with a terminal split with <leader-T>

Replace the makeprg command with <leader-M>
Run the makeprg command with <leader-m>

terminal.lua is needed to run the command that runs
the makeprg command in a TERMINAL. This allows
for interactive code in the terminal while running

However, if we find a match for errorformat, we migrate
the stdout of the terminal to the quickfix list and automatically
open the quickfix list. Getting the best of both worlds! (hopefully)